policies in job controller

+++ app/Http/Controllers/JobController.php
@@ -4,13 +4,15 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\View\View;
-use App\Models\Job;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Job;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -87,6 +89,8 @@ class JobController extends Controller
      */
     public function edit(Job $job): View
     {
+        // Check if user is authorized
+        $this->authorize("update", $job);
         return view("jobs.edit")->with("job", $job);
     }
 
@@ -95,6 +99,9 @@ class JobController extends Controller
      */
     public function update(Request $request, Job $job): string
     {
+        // Check if user is authorized
+        $this->authorize("update", $job);
+
         $validatedData = $request->validate([
             "title" => "required|string|max:255",
             "description" => "required|string",
@@ -139,6 +146,9 @@ class JobController extends Controller
      */
     public function destroy(Job $job): RedirectResponse
     {
+        // Check if user is authorized
+        $this->authorize("delete", $job);
+
         // If Logo, then delete it
         if ($job->company_logo) {
             Storage::delete("public/logos" . $job->company_logo);
